富有 form表单测试提交，还没通过

+++ com.jebao.root/1.1-com.jebao.p2p.web/src/main/java/com/jebao/p2p/web/utils/session/CurrentUser.java
@@ -4,15 +4,15 @@ package com.jebao.p2p.web.utils.session;
  * Created by Administrator on 2016/9/22.
  */
 public class CurrentUser {
-    private int Id;
+    private long Id;
     private String Name;
     private String UUID;
 
-    public int getId() {
+    public long getId() {
         return Id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         Id = id;
     }
     public String getName() {
